Code Review Updates, HomePOC refactor to Home, Other Changes
