--- .vscode/launch.json
@@ -1,22 +1,0 @@
-{
-    // Use IntelliSense to learn about possible attributes.
-    // Hover to view descriptions of existing attributes.
-    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-    "version": "0.2.0",
-    "configurations": [        
-        {
-            "type": "node",
-            "request": "launch",
-            "protocol": "legacy",
-            "name": "Start React",
-            "runtimeExecutable": "npm",
-            "runtimeArgs": [
-                "--prefix",
-                "hackru-frontend",
-                "start"
-            ],
-            "console": "integratedTerminal",
-            "port": 3000
-        }
-    ]
-}
