+++ hackru-frontend/public/index.html
@@ -9,7 +9,7 @@
     .##.....##.##.....##..######..##....##.##.....##..#######......######...######..#######.
 
     ©2019 HackRU Team at Rutgers University
-    Frontend: HackRU Marketing and Design Team (MnD)
+    Frontend: HackRU Frontend Team
     Backend: HackRU Research and Development Team (RnD)
 -->
 <html lang="en">
@@ -17,27 +17,14 @@
 <head>
     <meta charset="utf-8">
     <meta property="og:url" content="https://www.hackru.org/" />
-    <meta property="og:title" content="HackRU Fall 2018" />
-    <meta property="og:description" content="Get hyped for the best hackathon that Rutgers has to offer! The 13th semiannual HackRU is coming this Fall 2018, from October 6th-7th at College Ave Student Center." />
+    <meta property="og:title" content="HackRU Spring 2019" />
+    <meta property="og:description" content="Get hyped for the best hackathon that Rutgers has to offer! The 14th semiannual HackRU is coming this Spring 2019, from March 9th-10th at College Ave Student Center." />
     <meta property="og:image" content="./assets/fb-banner-upper.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="description" content="Sign up for HackRU Spring 2019, taking place from March 9th to 10th.">
     <link rel="shortcut icon" href="%PUBLIC_URL%/assets/fav.png">
-    <!--
-        manifest.json provides metadata used when your web app is added to the
-        homescreen on Android. See https://developers.google.com/web/fundamentals/web-app-manifest/
-    -->
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json">
     <link rel="stylesheet" href="%PUBLIC_URL%/index.css">
-    <!--
-        Notice the use of %PUBLIC_URL% in the tags above.
-        It will be replaced with the URL of the `public` folder during the build.
-        Only files inside the `public` folder can be referenced from the HTML.
-
-        Unlike "/favicon.ico" or "favicon.ico", "%PUBLIC_URL%/favicon.ico" will
-        work correctly both with client-side routing and a non-root public URL.
-        Learn how to configure a non-root public URL by running `npm run build`.
-    -->
     <title id="title">HackRU</title>
 </head>
 
@@ -46,16 +33,6 @@
         You need to enable JavaScript to view this website
     </noscript>
     <div id="root"></div>
-    <!--
-        This HTML file is a template.
-        If you open it directly in the browser, you will see an empty page.
-
-        You can add webfonts, meta tags, or analytics to this file.
-        The build step will place the bundled scripts into the <body> tag.
-
-        To begin the development, run `npm start` or `yarn start`.
-        To create a production bundle, use `npm run build` or `yarn build`.
-    -->
 </body>
 
 </html>
