+++ hackru-frontend/src/components/Landing/Landing.jsx
@@ -7,13 +7,11 @@
 /***************************************************************IMPORTS***************************************************************/
 import React, { Component } from "react";
 import { Container, Row } from "reactstrap";
-import Navigation from "./Navigation";
-import Home from "./Sections/Home";
 import Footer from "./Sections/Footer"
-import HomePOC from "./Sections/HomePOC";
-import Aliens from "./Aliens";
+import Home from "./Sections/Home";
+import FlyingLogo from "./FlyingLogo";
 import { ParallaxProvider } from "react-scroll-parallax";
-import { defaults, navlinks, theme } from "../../Defaults";
+import { navlinks, theme } from "../../Defaults";
 import ScrollableAnchor from "react-scrollable-anchor";
 /***************************************************************IMPORTS***************************************************************/
 
@@ -35,76 +33,46 @@ class LandingPage extends Component {
             if (navlinks[keys[i]].enabled) {
                 let url = navlinks[keys[i]].url.substring(1);
                 let component = navlinks[keys[i]].component;
+                let className = "bg-no-gradient skew-right";
                 if (i % 2 === 0) {
-                    rows.push((
-                        <ScrollableAnchor key={url} id={url}>
-                            <div>
-                                <Row style={{ ...sectionStyle }}> 
-                                    <div className="bg-gradient-right skew-right" >
-                                        <div className={sectionClasses}>
-                                            {component}
-                                        </div>
+                    className = "bg-gradient-right skew-right";
+                }
+                rows.push((
+                    <ScrollableAnchor key={url} id={url}>
+                        <div>
+                            <Row style={{ ...sectionStyle }}>
+                                <div className={className} >
+                                    <div className={sectionClasses}>
+                                        {component}
                                     </div>
-                                </Row>
-                            </div>
-                        </ScrollableAnchor>
-                    ));
-                } else {
-                    rows.push((
-                        <ScrollableAnchor key={url} id={url}>
-                            <div>
-                                <Row style={{ ...sectionStyle }}> 
-                                    <div className="bg-no-gradient skew-right" >
-                                        <div className={sectionClasses}>
-                                            {component}
-                                        </div>
-                                    </div>
-                                </Row> 
-                            </div>
-                        </ScrollableAnchor>
-                    ));
-                } 
+                                </div>
+                            </Row>
+                        </div>
+                    </ScrollableAnchor>
+                ));
             }
         }
-        if (!defaults.poc) {
-            return (
-                <Container id="LandingPage" fluid>
-                    <Navigation />
-                    <Row id="section" style={{ ...sectionStyle, background: "url(./assets/splash-bg-lowglow.png)", backgroundSize: "cover" }}>   
-                        <Home />   
-                    </Row>
-                    {rows}
-                </Container>
-            );
-        } else {
-            return (
-                <Container id="LandingPage" fluid style={{ ...sectionStyle, backgroundColor: theme.secondary[1] }}>
-                    <ParallaxProvider>
+        return (
+            <Container id="LandingPage" fluid style={{ ...sectionStyle, backgroundColor: theme.secondary[1] }}>
+                <ParallaxProvider>
                     <div style={{ position: "fixed", zIndex: 1, width: "100%", height: "100%", left: 0, top: 0, opacity: 0.5 }}>
-                        <Aliens />
+                        <FlyingLogo url={"./assets/icons/hru-alien-noplat-white.png"} />
                     </div>
-                    <div style={{ position: "fixed", zIndex: 2, width: "100%", height: "100%", left: 0, top: 0, background: "url(./assets/background.png)", backgroundSize: "cover", opacity: 0.5 }}>
-
-                    </div>
+                    <div style={{ position: "fixed", zIndex: 2, width: "100%", height: "100%", left: 0, top: 0, background: "url(./assets/background.png)", backgroundSize: "cover", opacity: 0.5 }}></div>
                     <ScrollableAnchor id="home">
                         <div>
                             <Row id="section" style={{ ...sectionStyle }}>
-                                <HomePOC />
+                                <Home />
                             </Row>
                         </div>
                     </ScrollableAnchor>
                     {rows}
-                    </ParallaxProvider>
-                    <div>
-                       
-                                <Footer />
-                            
-                       
-                        
-                    </div>
-                </Container>
-            );
-        }
+                </ParallaxProvider>
+                <div>
+                    <Footer />
+                </div>
+            </Container>
+        );
     }
 }
 /*****************************************************************APP*****************************************************************/
