+++ hackru-frontend/src/components/Landing/Sections/Sponsors.jsx
@@ -17,7 +17,7 @@ class Sponsors extends Component {
         return (
             <div>
                 <h1 className="display-4 theme-font mb-3">Sponsorship</h1>
-                <p className=""> Our Spring 2019 Sponsors list is coming soon.  In the meantime, if you are interested in sponsoring HackRU, please email <a href="emailto:sponsorship@hackru.org">sponsorship@hackru.org</a>.</p>
+                <p className="lead"> Our Spring 2019 Sponsors list is coming soon.  In the meantime, if you are interested in sponsoring HackRU, please email <a href="emailto:sponsorship@hackru.org">sponsorship@hackru.org</a>.</p>
             </div>
         )
     }
