+++ hackru-frontend/src/index.js
@@ -35,5 +35,3 @@ function main(worker) {
 }
 main(false)
 /*****************************************************************APP*****************************************************************/
-
-// Kill me now ~ Vijay Modha @8:13pm,12/9/18
